requirements.txt
Pillow>=10.4.0
numpy>=2.1.1
scikit-learn>=1.5.2
opencv-python>=4.10.0